+++ ameriflux_pipeline/pre_pyfluxpro.py
@@ -88,10 +88,10 @@ def eddypro_preprocessing(file_meta_data_file):
         log.error("Creation of master met data has failed.")
         return None
     # write processed df to output path
-    data_util.write_data(df, cfg.MASTER_MET)
+    data_util.write_data_to_csv(df, cfg.MASTER_MET)
 
     # Write file meta data to another file
-    data_util.write_data(file_meta, file_meta_data_file)  # write meta data of file to file. One row.
+    data_util.write_data_to_csv(file_meta, file_meta_data_file)  # write meta data of file to file. One row.
 
     # create file for master met formatted for eddypro
     # filename is selected to be master_met_eddypro
@@ -104,7 +104,7 @@ def eddypro_preprocessing(file_meta_data_file):
         log.error("Eddypro formatting of master met data failed.")
         return None
     # write formatted df to output path
-    data_util.write_data(df, eddypro_formatted_met_file)
+    data_util.write_data_to_csv(df, eddypro_formatted_met_file)
 
     return eddypro_formatted_met_file
 
@@ -147,11 +147,11 @@ def pyfluxpro_processing(eddypro_full_output, full_output_pyfluxpro, met_data_30
     full_output_df['TIMESTAMP'][1:] = pd.to_datetime(full_output_df['TIMESTAMP'][1:])
 
     # write pyfluxpro formatted df to output path
-    data_util.write_data(full_output_df, full_output_pyfluxpro)
+    data_util.write_data_to_csv(full_output_df, full_output_pyfluxpro)
     # copy and rename the met data file
     shutil.copyfile(met_data_30_input, met_data_30_pyfluxpro)
 
-    met_data_df = pd.read_csv(met_data_30_pyfluxpro, low_memory=False)
+    met_data_df = data_util.read_csv_file(met_data_30_pyfluxpro, dtype='unicode')
     # convert timestamp to datetime format so that pyfluxpro can read without error
     met_data_df['TIMESTAMP'][1:] = pd.to_datetime(met_data_df['TIMESTAMP'][1:])
 
@@ -344,11 +344,11 @@ def pre_processing(file_meta_data_file, erroring_variable_flag):
 
     # if eddypro full output file not present
     if not eddypro_full_outfile:
-        log.error("EddyPro full output not present. Aborting")
+        log.error('-' * 10 + "EddyPro full output not present. Aborting" + '-' * 10)
         # return failure
         return False
     if not is_pyfluxpro_processing_success:
-        log.error("PyFluxpro processing failed. Aborting")
+        log.error('-' * 10 + "PyFluxpro processing failed. Aborting" + '-' * 10)
         return False
 
     # run ameriflux formatting of pyfluxpro input
@@ -361,7 +361,7 @@ def pre_processing(file_meta_data_file, erroring_variable_flag):
         # return failure
         return False
     if not is_pyfluxpro_ameriflux_processing_success:
-        log.error("PyFluxpro input sheet formatting for Ameriflux failed. Aborting")
+        log.error('-' * 10 + "PyFluxpro input sheet formatting for Ameriflux failed. Aborting" + '-' * 10)
         return False
 
     # run ameriflux formatting of pyfluxpro L1 control file
@@ -372,7 +372,7 @@ def pre_processing(file_meta_data_file, erroring_variable_flag):
                                           cfg.L1_AMERIFLUX, erroring_variable_flag,
                                           cfg.L1_AMERIFLUX_ERRORING_VARIABLES_KEY)
     if pyfluxpro_ameriflux_labels is None:
-        log.error("PyFluxPro L1 processing failed. Aborting")
+        log.error('-' * 10 + "PyFluxPro L1 processing failed. Aborting" + '-' * 10)
         # return failure
         return False
 
@@ -386,7 +386,7 @@ def pre_processing(file_meta_data_file, erroring_variable_flag):
         # all processing done return success
         return True
     else:
-        log.error("PyFluxPro L2 processing failed. Aborting")
+        log.error('-' * 10 + "PyFluxPro L2 processing failed. Aborting" + '-' * 10)
         # return failure
         return False
 
@@ -400,7 +400,7 @@ def main():
     # Main function
     is_valid_config = input_validation()
     if not is_valid_config:
-        log.error("Check .env file and fix configurations. Aborting")
+        log.error('-' * 10 + "Check .env file and fix configurations. Aborting" + '-' * 10)
         return
 
     # Some preprocessing
@@ -430,7 +430,7 @@ def main():
     if is_success:
         log.info("Successfully completed pre-processing of PyFluxPro L1 and L2")
     else:
-        log.error("Pre-processing resulted in an error.")
+        log.error('-' * 10 + "Pre-processing resulted in an error." + '-' * 10)
 
     end = time.time()
     hours, rem = divmod(end - start, 3600)