+++ ameriflux_pipeline/pyfluxpro/l2format.py
@@ -9,6 +9,7 @@ import os
 import re
 import logging
 
+import utils.data_util as data_util
 from utils.process_validation import DataValidation, L2Validation
 
 # create log object with current module name
@@ -128,8 +129,9 @@ class L2Format:
         l2_output_lines.extend(plot_lines_out)
 
         # write output lines to file
-        L2Format.write_list_to_file(l2_output_lines, l2_ameriflux_output)
-        # processing completed
+        log.info("Writting Ameriflux L2 control file to " + l2_ameriflux_output)
+        data_util.write_list_to_file(l2_output_lines, l2_ameriflux_output)
+        # process successfully completed
         return True
 
     @staticmethod
@@ -176,6 +178,7 @@ class L2Format:
         """
         variables = text[text.str.contains(var_pattern)]
         var_start_end = []
+        end_ind = text.first_valid_index()  # initialize end index
         for i in range(len(variables.index) - 1):
             start_ind = variables.index[i]
             end_ind = variables.index[i + 1]
@@ -397,20 +400,3 @@ class L2Format:
                 plot_lines_out[ind] = line
 
         return plot_lines_out
-
-    @staticmethod
-    def write_list_to_file(in_list, outfile):
-        """
-            Save list with string to a file
-            Args:
-                in_list (list): List of the strings
-                outfile (str): A file path of the output file
-            Returns:
-                None
-        """
-        try:
-            with open(outfile, 'w') as f:
-                f.write('\n'.join(in_list))
-            log.info("AmeriFlux L2 saved in %s", outfile)
-        except Exception as e:
-            log.error("Failed to create file %s. %s", outfile, e)